Fix plugin not running on example imports, extract color functions to another file
